+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.20.0] - 2026-01-07
+
+### Added
+- New "Investors" field for organizations allowing both people and companies to be listed as investors
+- Investors section displayed on organization detail page with links to people/companies
+- Multi-select investor picker in organization form with search
+
 ## [1.19.3] - 2026-01-07
 
 ### Added